Correcting mouse-positioning issue from gnome-shell forcing window-resizing.
The skia canvas is sized to the initial / full-screen size, but as the SDL window is created
and hits the border of gnome-shell, the windows manager forces the SDL window to resize to fit.
This generates a windows-size changed event, and can be detected.

The mouse position is monitored as related to the current windows size, but is drawn
to the skia canvas assuming the initial windows size. The difference is about 100 pixels
from gnome-shell's top and bottom menu, plus the SDL window's own menu bar.

Initially filed as https://github.com/HinTak/skia-c-examples/issues/2 .

Fixes #3