Fixes crash, and other diagnostic improvements/additions
In the end of day, it is swapping
    canvas.drawImage(self.star_image, -50.0, -50.0)
with
    canvas.drawPath(create_star(), self.paint)
which fixes the crash. Pointer indicating that the grcontext might be
invalid, etc, but it seems more like it is cpuSurface vs GL canvas.
Caching both cpuSurface and offscreen does not work.

Comment out the drawImage line avoid the crash, but can result in
"Could not allocate vertices" in stderr.

While it works now, and works better as an application,
it is less of a demo for mixed skia cpu/gpu api with Qt integration.

Curiously, Copilot made the image->path switch in the Qt c port.

To investigate. Perhaps by modifying the c code to ->image, and see why
it breaks, as well as try this under pyQt5.